fix(provisioning): split instrumentation so Edge bundle skips mysql2
CI build failed because Next.js bundles instrumentation.ts for both Node
and Edge runtimes. Even though our register() function gated all work
behind `process.env.NEXT_RUNTIME === 'nodejs'`, webpack still resolved
the transitive imports for the Edge bundle, which has no `fs`/`path`/
`net`/`tls`/`timers`.

Standard Next.js fix: keep instrumentation.ts as a thin shim and move
all Node-only work to instrumentation-node.ts. The dynamic
`import('./instrumentation-node')` is only reached in the Node bundle.
