+++ frontend/instrumentation.ts
@@ -1,43 +1,12 @@
 /**
- * Next.js instrumentation hook. Runs once per Node.js process at startup.
- *
- * Responsibilities:
- *   - Install SIGTERM/SIGINT handlers in the provisioning worker so a graceful
- *     shutdown stops heartbeats (rows go stale, next process recovers them).
- *   - Pick up any `running` provisioning runs whose worker heartbeat went
- *     silent — these are orphans from a previous process that crashed or was
- *     SIGKILLed.
- *
- * Edge runtime: the hook also runs in the Edge runtime; we skip there because
- * mysql2 is Node-only.
+ * Next.js instrumentation hook. Runs once per process at startup, in BOTH the
+ * Node and Edge runtimes. We only need to install provisioning hooks in Node,
+ * and the Node-only work transitively imports mysql2 which has no Edge build —
+ * so the dynamic import of the Node implementation must be gated by runtime
+ * and live in a separate file that the Edge bundle never resolves.
  */
 export async function register() {
-  if (process.env.NEXT_RUNTIME !== 'nodejs') return;
-
-  let ailogger: typeof import('@/ailogger').default | null = null;
-  try {
-    const aiModule = await import('@/ailogger');
-    ailogger = aiModule.default;
-  } catch {
-    // ailogger init can fail in early bootstrap (e.g., AppInsights not ready);
-    // proceed without it so worker pickup can still run.
-  }
-
-  try {
-    const { pickupStaleRuns, installShutdownHandler } = await import('@/lib/provisioning/worker');
-    const { getPoolMonitorInstance } = await import('@/config/poolmonitorsingleton');
-    const pool = getPoolMonitorInstance().pool;
-    installShutdownHandler(pool);
-    const picked = await pickupStaleRuns(pool);
-    if (picked.length > 0 && ailogger) {
-      ailogger.info('provisioning.worker.startup', { pickedUpRuns: picked });
-    }
-  } catch (err) {
-    const errorMessage = err instanceof Error ? err.message : String(err);
-    if (ailogger) {
-      ailogger.warn('provisioning.worker.startup_failed', { errorMessage });
-    } else {
-      console.warn('[instrumentation] provisioning worker startup failed:', errorMessage);
-    }
+  if (process.env.NEXT_RUNTIME === 'nodejs') {
+    await import('./instrumentation-node');
   }
 }
